+++ tasks/seq/oturin_a_max_values_by_rows_matrix_seq/src/ops_seq.cpp
@@ -1,5 +1,6 @@
 #include "seq/oturin_a_max_values_by_rows_matrix_seq/include/ops_seq.hpp"
 
+#include <algorithm>
 #include <thread>
 
 using namespace std::chrono_literals;
@@ -7,8 +8,8 @@ using namespace std::chrono_literals;
 bool oturin_a_max_values_by_rows_matrix_seq::TestTaskSequential::pre_processing() {
   internal_order_test();
   // Init vectors
-  n = (int)(taskData->inputs_count[0] / m);
-  m = (int)*taskData->inputs[1];
+  n = (size_t)(taskData->inputs_count[0] / m);
+  m = (size_t)*taskData->inputs[1];
   input_ = std::vector<int>(taskData->inputs_count[0]);
   int *tmp_ptr = reinterpret_cast<int *>(taskData->inputs[0]);
   for (unsigned i = 0; i < taskData->inputs_count[0]; i++) {
@@ -27,13 +28,13 @@ bool oturin_a_max_values_by_rows_matrix_seq::TestTaskSequential::validation() {
 
 bool oturin_a_max_values_by_rows_matrix_seq::TestTaskSequential::run() {
   internal_order_test();
-  for (int i = 0; i < m; i++) res[i] = *std::max_element(input_.begin() + i * n, input_.begin() + (i + 1) * n);
+  for (size_t i = 0; i < m; i++) res[i] = *std::max_element(input_.begin() + i * n, input_.begin() + (i + 1) * n);
   return true;
 }
 
 bool oturin_a_max_values_by_rows_matrix_seq::TestTaskSequential::post_processing() {
   internal_order_test();
-  for (int i = 0; i < m; i++) {
+  for (size_t i = 0; i < m; i++) {
     reinterpret_cast<int *>(taskData->outputs[0])[i] = res[i];
   }
   return true;
